Create DATA310-monte-carlo
Monte Carlo Simulations in R for Probability and Statistics:

This project involved utilizing Monte Carlo simulations in R to explore various probability distributions and statistical problems. The work included calculating probabilities for normal distributions, constructing confidence intervals, and simulating dice rolls and coin flips. In one task, R code was modified to approximate the probability of specific outcomes in a Yahtzee game, such as rolling a "large straight." Another task involved simulating the longest streak of heads or tails in a series of coin flips, allowing for the creation of a cumulative distribution function (CDF) to determine the likelihood of specific streak lengths. The project demonstrated proficiency in statistical concepts, probability theory, and R programming, particularly in the context of simulations and probabilistic modeling.